chore(testworkflows): TKC-1763 - cleanup ServiceState
